added property based testing
